// src/types/openapi.d.ts
export interface components {
    schemas: {
// ...
        Page: {
            /**
             * @description ID of the page.
             * @example canvas-IjkLmnO
             */
            id: string;
            /**
             * @description The type of this resource.
             * @enum {string}
             */
            type: "page";
            /**
             * Format: url
             * @description API link to the page.
             * @example https://coda.io/apis/v1/docs/AbCDeFGH/pages/canvas-IjkLmnO
             */
            href: string;
            /**
             * Format: url
             * @description Browser-friendly link to the page.
             * @example https://coda.io/d/_dAbCDeFGH/Launch-Status_sumnO
             */
            browserLink: string;
            /**
             * @description Name of the page.
             * @example Launch Status
             */
            name: string;
            /**
             * @description Subtitle of the page.
             * @example See the status of launch-related tasks.
             */
            subtitle?: string;
            icon?: components["schemas"]["Icon"];
            image?: components["schemas"]["Image"];
            contentType: components["schemas"]["PageType"];
            /**
             * @description Whether the page is hidden in the UI.
             * @example true
             */
            isHidden: boolean;
            /**
             * @description Whether the page or any of its parents is hidden in the UI.
             * @example true
             */
            isEffectivelyHidden: boolean;
            parent?: components["schemas"]["PageReference"];
            children: components["schemas"]["PageReference"][];
            /** @description Authors of the page */
            authors?: components["schemas"]["PersonValue"][];
            /**
             * Format: date-time
             * @description Timestamp for when the page was created.
             * @example 2018-04-11T00:18:57.946Z
             */
            createdAt?: string;
            createdBy?: components["schemas"]["PersonValue"];
            /**
             * Format: date-time
             * @description Timestamp for when page content was last modified.
             * @example 2018-04-11T00:18:57.946Z
             */
            updatedAt?: string;
            updatedBy?: components["schemas"]["PersonValue"];
        };
        /** @description List of pages. */
        PageList: {
            items: components["schemas"]["Page"][];
            /**
             * Format: url
             * @description API link to these results
             * @example https://coda.io/apis/v1/docs/AbCDeFGH/pages?limit=20
             */
            href?: string;
            nextPageToken?: components["schemas"]["nextPageToken"];
            nextPageLink?: components["schemas"]["nextPageLink"] & string;
        };
// ...
        PageUpdate: {
            /**
             * @description Name of the page.
             * @example Launch Status
             */
            name?: string;
            /**
             * @description Subtitle of the page.
             * @example See the status of launch-related tasks.
             */
            subtitle?: string;
            /**
             * @description Name of the icon.
             * @example rocket
             */
            iconName?: string;
            /**
             * @description Url of the cover image to use.
             * @example https://example.com/image.jpg
             */
            imageUrl?: string;
            /**
             * @description Whether the page is hidden or not. Note that for pages that cannot be hidden, like the sole top-level page in a doc, this will be ignored.
             * @example true
             */
            isHidden?: boolean;
            contentUpdate?: Record<string, never> & components["schemas"]["PageContentUpdate"];
        };
// ...
        BeginPageContentExportRequest: {
            outputFormat: components["schemas"]["PageContentOutputFormat"];
        };
// ...
    };
// ...
}